% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boursorama_connect.R
\name{connect_boursorama}
\alias{connect_boursorama}
\title{connect_boursorama}
\usage{
connect_boursorama(account_number = NULL, password = NULL)
}
\arguments{
\item{account_number}{an integer corresponding to the cient account number}

\item{password}{an integer corresponding to the password which connect to the account}
}
\description{
connect to the boursorama bank webpage
}

