% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_qif.R
\name{read_qif}
\alias{read_qif}
\title{read_qif}
\arguments{
\item{file}{the name of the file which the data are to be read from. Each recording operation is separated by a blank line in the file. File is read from \code{\link{read.table}}}
}
\value{
A data frame containing a representation of the data in the file.
}
\description{
Reads a file in qif format and creates a data frame from it, with operations corresponding to lines and variables to fields in the file.
}

